--- packages/dart/lib/src/network/http_client_adapter.dart
@@ -1,2 +1,0 @@
-export 'http_client_adapter_native.dart'
-    if (dart.library.js) 'http_client_adapter_web.dart';
